requirements.txt
langchain
langchain-community
langchain-core
langchain-google-genai
langchain-huggingface
faiss-cpu
streamlit
pypdf
python-dotenv
tiktoken
sentence-transformers
